+++ src/inspector/models/measures/measuressettingsmodel.h
@@ -39,7 +39,7 @@ class MeasuresSettingsModel : public AbstractInspectorModel
     Q_PROPERTY(bool allSystemsAreLocked READ allSystemsAreLocked NOTIFY allSystemsAreLockedChanged)
     Q_PROPERTY(bool scoreIsInPageView READ scoreIsInPageView NOTIFY scoreIsInPageViewChanged)
     Q_PROPERTY(bool isMakeIntoSystemAvailable READ isMakeIntoSystemAvailable NOTIFY isMakeIntoSystemAvailableChanged)
-    Q_PROPERTY(int systemCount READ systemCount NOTIFY systemCountChanged)
+    Q_PROPERTY(size_t systemCount READ systemCount NOTIFY systemCountChanged)
 
 public:
     explicit MeasuresSettingsModel(QObject* parent, IElementRepositoryService* repository);
@@ -79,7 +79,7 @@ public:
     bool scoreIsInPageView() const;
     bool isMakeIntoSystemAvailable() const;
 
-    int systemCount() const;
+    size_t systemCount() const;
 
 protected:
     void onNotationChanged(const mu::engraving::PropertyIdSet&, const mu::engraving::StyleIdSet&) override;
@@ -94,13 +94,13 @@ signals:
     void allSystemsAreLockedChanged(bool allLocked);
     void scoreIsInPageViewChanged(bool isInPageView);
     void isMakeIntoSystemAvailableChanged(bool isMakeIntoSystemAvailable);
-    void systemCountChanged(int count);
+    void systemCountChanged(size_t count);
 
 private:
     bool m_allSystemsAreLocked = false;
     bool m_scoreIsInPageView = false;
     bool m_isMakeIntoSystemAvailable = false;
-    int m_systemCount = 0;
+    size_t m_systemCount = 0;
 };
 }
